fix object types

+++ src/encoding/abstract.jl
@@ -5,12 +5,12 @@ Base type for state encoders that transform simulation state into
 numerical representations suitable for neural networks.
 
 All concrete encoders must implement:
-- `encode_state(encoder, sim, ambulance_index) -> Vector{Float64}`
+- `encode_state(encoder, sim, ambulance_index) -> Vector{Float32}`
 """
 abstract type AbstractEncoder end
 
 """
-    encode_state(encoder::AbstractEncoder, sim::JEMSS.Simulation, ambulance_index::Int) -> Vector{Float64}
+    encode_state(encoder::AbstractEncoder, sim::JEMSS.Simulation, ambulance_index::Int) -> Vector{Float32}
 
 Transform the current simulation state into a numerical vector representation.
 
@@ -20,9 +20,9 @@ Transform the current simulation state into a numerical vector representation.
 - `ambulance_index`: Index of the triggering ambulance
 
 # Returns
-- `Vector{Float64}`: Encoded state suitable for neural network input
+- `Vector{Float32}`: Encoded state suitable for neural network input
 """
-function encode_state(encoder::AbstractEncoder, sim::JEMSS.Simulation, ambulance_index::Int)::Vector{Float64}
+function encode_state(encoder::AbstractEncoder, sim::JEMSS.Simulation, ambulance_index::Int)::Vector{Float32}
     error("encode_state not implemented for $(typeof(encoder))")
 end
